feat: refactor matching API with MatchMode bitmask and remove createGit
- Remove `createGit` factory function from `Git` target.
- Replace `nocase` option and `re.test` lower argument with `MatchMode` bitmask.
- Introduce `MatchMode` flags: `normal`, `unsensitive`, `wildmatch`, `lowered`.
- Implement `wildmatch` mode with `noextglob: true` and `nobrace: true`.
- Optimize case-insensitive matching by avoiding redundant `toLowerCase()` via `MatchMode.lowered`.
- Refactor all targets and extraction modules to the new bitwise API.
- Solve circular dependencies by extracting `MatchMode` and `PatternCache` types to `patternMode.ts`.
- Update tests and benchmarks to verify performance and correctness.
