+++ src/patterns/gitignore.ts
@@ -1,7 +1,7 @@
 import type { ExtractorFn } from "./extractor.js"
 import type { Rule } from "./rule.js"
 
-import { MatchMode } from "./patternList.js"
+import { MatchMode } from "./patternMode.js"
 import { ruleCompile } from "./resolveSources.js"
 import { resolveNegatable, type Source } from "./source.js"
