+++ src/patterns/index.ts
@@ -8,7 +8,7 @@ export * from "./matcherContext.js"
 export * from "./matcherContextPatch.js"
 export * from "./matcherStream.js"
 export * from "./packagejson.js"
-export * from "./patternList.js"
+export * from "./patternMode.js"
 export * from "./resolveSources.js"
 export * from "./resource.js"
 export * from "./rule.js"
