+++ src/patterns/patternCompile.ts
@@ -1,6 +1,26 @@
 import glob from "micromatch"
 
-import { MatchMode, type PatternCache, type PatternList } from "./patternList.js"
+import { MatchMode, type PatternCache, type PatternList } from "./patternMode.js"
+
+/**
+ * Compiles the {@link PatternList}.
+ *
+ * @see {@link patternCompile}
+ * @see {@link ruleCompile}
+ *
+ * @since 0.6.0
+ */
+export function patternListCompile(
+	list: PatternList,
+	mode: MatchMode = MatchMode.normal,
+): PatternCache[] {
+	const len = list.length
+	const res = Array.from<PatternCache>({ length: len })
+	for (let i = 0; i < len; i++) {
+		res[i] = patternCompile(list[i]!, list, mode)
+	}
+	return res
+}
 
 /**
  * Compiles a string of the {@link PatternList}.
