+++ src/targets/git.ts
@@ -136,4 +136,3 @@ export const Git: Target = <Target>{
 	internalRules: defIntRules,
 	root: "/",
 }
-
